requirements.txt
FastAPI==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0              # PostgreSQL driver for async SQLAlchemy
psycopg2-binary==2.9.9       # PostgreSQL driver for Alembic/sync operations
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1     # For managing settings from env vars
python-dotenv==1.0.1
redis==5.0.4
celery==5.3.6
PyMuPDF==1.23.26             # For PDF extraction (fitz)
langchain==0.1.16            # Or specific LangChain components if not using custom
langchain-community==0.0.33  # For vector stores, document loaders etc.
chromadb==0.4.24             # Or faiss-cpu/faiss-gpu
google-auth-library==9.9.0   # For Google ID token verification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0                # For making HTTP requests to external APIs
openai==1.25.1               # Required for inference.py LLM evaluation wrapper
loguru==0.7.2                # For structured logging
python-multipart==0.0.9

# Test dependencies
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.12.0
